UIAPPS-69 Allow any example to be installed
We also document a bit how this command works.

Since we're downloading an archive, and selecting files from it, we
don't have to limit ourselves to just the `starter-kit`. We can actually
use any example in the `examples` directory. There aren't a whole lot of
examples at the moment, but this should allow us to expand them
relatively easily.