UIAPPS-69 Update documentation to use init package
This should be the way to go going forward.

As nice as it would be to also do the `cd` as part of the package, that
doesn't seem to be possible on a POSIX level. A child process cannot
easily change the working directory of a parent process:
https://stackoverflow.com/a/37624361. Given that we're running in a
Node.js process and we expect people to have called this from some sort
of shell/command line, there's just not an easy way to actually affect
the parent process. Not the end of the world, but something worth
noting.

We are installing the dependencies, however, so we get rid of that line
in the instructions. One less thing for App developers to worry about.

We don't mention any of the `npm` workflow here, since we don't mention
`npm` anywhere else in any of the documentation. But, we could also do
that if we wanted to make it a bit more agnostic.